tests/test_services.py
"""
Unit tests for E-commerce microservices.
"""

import pytest
from unittest.mock import AsyncMock, patch
from datetime import datetime


class TestUserService:
    """Tests for User Service."""
    
    def test_password_hashing(self):
        """Test password hashing and verification."""
        from passlib.context import CryptContext
        pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
        
        password = "test_password_123"
        hashed = pwd_context.hash(password)
        
        assert pwd_context.verify(password, hashed)
        assert not pwd_context.verify("wrong_password", hashed)
    
    def test_jwt_token_creation(self):
        """Test JWT token creation and decoding."""
        from jose import jwt
        
        secret = "test-secret"
        algorithm = "HS256"
        user_id = "user-123"
        
        token = jwt.encode({"sub": user_id}, secret, algorithm=algorithm)
        decoded = jwt.decode(token, secret, algorithms=[algorithm])
        
        assert decoded["sub"] == user_id


class TestOrderService:
    """Tests for Order Service."""
    
    def test_order_total_calculation(self):
        """Test order total calculation."""
        items = [
            {"quantity": 2, "unit_price": 10.00},
            {"quantity": 1, "unit_price": 25.50},
            {"quantity": 3, "unit_price": 5.00}
        ]
        
        total = sum(item["quantity"] * item["unit_price"] for item in items)
        
        assert total == 60.50
    
    def test_order_status_transitions(self):
        """Test valid order status transitions."""
        valid_transitions = {
            "pending": ["confirmed", "cancelled"],
            "confirmed": ["processing", "cancelled"],
            "processing": ["shipped"],
            "shipped": ["delivered"],
            "delivered": [],
            "cancelled": []
        }
        
        # Pending can go to confirmed
        assert "confirmed" in valid_transitions["pending"]
        
        # Delivered cannot transition
        assert len(valid_transitions["delivered"]) == 0


class TestInventoryService:
    """Tests for Inventory Service."""
    
    def test_available_quantity_calculation(self):
        """Test available quantity calculation."""
        stock_quantity = 100
        reserved_quantity = 25
        
        available = stock_quantity - reserved_quantity
        
        assert available == 75
    
    def test_stock_reservation(self):
        """Test stock reservation logic."""
        stock = {"quantity": 100, "reserved": 0}
        
        # Reserve 30 units
        reserve_amount = 30
        if stock["quantity"] - stock["reserved"] >= reserve_amount:
            stock["reserved"] += reserve_amount
        
        assert stock["reserved"] == 30
        assert stock["quantity"] - stock["reserved"] == 70
    
    def test_insufficient_stock(self):
        """Test insufficient stock detection."""
        stock = {"quantity": 10, "reserved": 5}
        available = stock["quantity"] - stock["reserved"]
        
        # Try to reserve 10 units
        reserve_amount = 10
        
        assert available < reserve_amount


class TestPaymentService:
    """Tests for Payment Service."""
    
    def test_payment_status_values(self):
        """Test payment status enum values."""
        valid_statuses = ["pending", "processing", "completed", "failed", "refunded"]
        
        assert "completed" in valid_statuses
        assert "refunded" in valid_statuses
    
    def test_payment_method_values(self):
        """Test payment method enum values."""
        valid_methods = ["credit_card", "debit_card", "paypal", "bank_transfer"]
        
        assert "credit_card" in valid_methods
        assert "paypal" in valid_methods


class TestEventTypes:
    """Tests for Kafka event types."""
    
    def test_event_type_format(self):
        """Test event type naming convention."""
        event_types = [
            "user.registered",
            "order.created",
            "order.confirmed",
            "inventory.reserved",
            "payment.completed"
        ]
        
        for event_type in event_types:
            parts = event_type.split(".")
            assert len(parts) == 2
            assert parts[0] in ["user", "order", "inventory", "payment"]
    
    def test_event_payload_structure(self):
        """Test event payload has required fields."""
        event = {
            "event_type": "order.created",
            "order_id": "order-123",
            "user_id": "user-456",
            "timestamp": datetime.utcnow().isoformat()
        }
        
        assert "event_type" in event
        assert "timestamp" in event


class TestCacheKeys:
    """Tests for Redis cache key patterns."""
    
    def test_user_cache_key(self):
        """Test user cache key format."""
        user_id = "user-123"
        key = f"user:{user_id}"
        
        assert key == "user:user-123"
    
    def test_product_cache_key(self):
        """Test product cache key format."""
        product_id = "prod-001"
        key = f"product:{product_id}"
        
        assert key == "product:prod-001"
    
    def test_order_cache_key(self):
        """Test order cache key format."""
        order_id = "order-789"
        key = f"order:{order_id}"
        
        assert key == "order:order-789"


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
