requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0
aioredis>=2.0.0

# Kafka
aiokafka>=0.10.0
kafka-python>=2.0.2

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Monitoring
prometheus-client>=0.19.0
